frontend/requirements.txt
streamlit
google-cloud-firestore
pandas
plotly
